--- components/TimeChartForSubsidyApplication.vue
@@ -1,551 +1,0 @@
-<template>
-  <data-view :title="title" :title-id="titleId" :date="date">
-    <template v-slot:description>
-      <slot name="description" />
-    </template>
-    <ul
-      :class="$style.GraphLegend"
-      :style="{ display: canvas ? 'block' : 'none' }"
-    >
-      <li v-for="(item, i) in items" :key="i" @click="onClickLegend(i)">
-        <button>
-          <div
-            :style="{
-              backgroundColor: colors[i].fillColor,
-              borderColor: colors[i].strokeColor
-            }"
-          />
-          <span
-            :style="{
-              textDecoration: displayLegends[i] ? 'none' : 'line-through'
-            }"
-            >{{ item }}</span
-          >
-        </button>
-      </li>
-    </ul>
-    <h4 :id="`${titleId}-graph`" class="visually-hidden">
-      {{ $t(`{title}のグラフ`, { title }) }}
-    </h4>
-    <div class="LegendStickyChart">
-      <div class="scrollable" :style="{ display: canvas ? 'block' : 'none' }">
-        <div :style="{ width: `${graphWidth}px` }">
-          <bar
-            :ref="'barChart'"
-            :chart-id="chartId"
-            :chart-data="displayData"
-            :options="displayOption"
-            :plugins="scrollPlugin"
-            :display-legends="displayLegends"
-            :height="240"
-            :width="graphWidth"
-          />
-        </div>
-      </div>
-      <bar
-        class="sticky-legend"
-        :style="{ display: canvas ? 'block' : 'none' }"
-        :chart-id="`${chartId}-header`"
-        :chart-data="displayDataHeader"
-        :options="displayOptionHeader"
-        :plugins="yAxesBgPlugin"
-        :display-legends="displayLegends"
-        :height="240"
-      />
-    </div>
-    <template v-slot:infoPanel>
-      <data-view-basic-info-panel
-        :l-text="displayInfo.lText"
-        :m-text="displayInfo.mText"
-        :s-text="displayInfo.sText"
-        :unit="displayInfo.unit"
-      />
-    </template>
-    <template v-slot:footer>
-      <open-data-link v-show="url" :url="url" />
-    </template>
-  </data-view>
-</template>
-
-<script lang="ts">
-import Vue from 'vue'
-import { ThisTypedComponentOptionsWithRecordProps } from 'vue/types/options'
-import { TranslateResult } from 'vue-i18n'
-import { Chart } from 'chart.js'
-import DataView from '@/components/DataView.vue'
-import DataSelector from '@/components/DataSelector.vue'
-import DataViewTable from '@/components/DataViewTable.vue'
-import DataViewBasicInfoPanel from '@/components/DataViewBasicInfoPanel.vue'
-import { DisplayData, yAxesBgPlugin, scrollPlugin } from '@/plugins/vue-chart'
-import { getGraphSeriesStyle, SurfaceStyle } from '@/utils/colors'
-import OpenDataLink from '@/components/OpenDataLink.vue'
-
-type Data = {
-  canvas: boolean
-  displayLegends: boolean[]
-  colors: SurfaceStyle[]
-  chartWidth: number | null
-}
-type Methods = {
-  sum: (array: number[]) => number
-  cumulative: (array: number[]) => number[]
-  pickLastNumber: (chartDataArray: number[][]) => number[]
-  cumulativeSum: (chartDataArray: number[][]) => number[]
-  eachArraySum: (chartDataArray: number[][]) => number[]
-  onClickLegend: (i: number) => void
-}
-
-type Computed = {
-  displayInfo: {
-    lText: string
-    mText: string
-    sText: TranslateResult
-    unit: string
-  }
-  displayData: DisplayData
-  displayOption: Chart.ChartOptions
-  displayDataHeader: DisplayData
-  displayOptionHeader: Chart.ChartOptions
-  scaledTicksYAxisMax: number
-  graphWidth: number
-}
-
-type Props = {
-  title: string
-  titleId: string
-  chartId: string
-  chartData: number[][]
-  date: string
-  items: string[]
-  labels: string[]
-  dataLabels: string[] | TranslateResult[]
-  tableLabels: string[] | TranslateResult[]
-  unit: string
-  scrollPlugin: Chart.PluginServiceRegistrationOptions[]
-  yAxesBgPlugin: Chart.PluginServiceRegistrationOptions[]
-  url: string
-}
-
-const options: ThisTypedComponentOptionsWithRecordProps<
-  Vue,
-  Data,
-  Methods,
-  Computed,
-  Props
-> = {
-  created() {
-    this.canvas = process.browser
-  },
-  components: {
-    DataView,
-    DataSelector,
-    DataViewTable,
-    DataViewBasicInfoPanel,
-    OpenDataLink
-  },
-  props: {
-    title: {
-      type: String,
-      default: ''
-    },
-    titleId: {
-      type: String,
-      required: false,
-      default: ''
-    },
-    chartId: {
-      type: String,
-      default: 'time-stacked-bar-chart'
-    },
-    chartData: {
-      type: Array,
-      required: false,
-      default: () => []
-    },
-    date: {
-      type: String,
-      required: true,
-      default: ''
-    },
-    items: {
-      type: Array,
-      default: () => []
-    },
-    labels: {
-      type: Array,
-      default: () => []
-    },
-    dataLabels: {
-      type: Array,
-      default: () => []
-    },
-    tableLabels: {
-      type: Array,
-      default: () => []
-    },
-    unit: {
-      type: String,
-      default: ''
-    },
-    scrollPlugin: {
-      type: Array,
-      default: () => scrollPlugin
-    },
-    yAxesBgPlugin: {
-      type: Array,
-      default: () => yAxesBgPlugin
-    },
-    url: {
-      type: String,
-      required: false,
-      default: ''
-    }
-  },
-  data: () => ({
-    displayLegends: [true, true],
-    colors: getGraphSeriesStyle(2),
-    chartWidth: null,
-    canvas: true
-  }),
-  computed: {
-    displayInfo() {
-      const len = this.chartData[0].length
-      const cnt = this.chartData[1][len - 1]
-      return {
-        lText: cnt.toLocaleString(),
-        mText: '累計値',
-        sText: this.$t('支給決定件数'),
-        unit: this.unit
-      }
-    },
-    displayData() {
-      const graphSeries = getGraphSeriesStyle(this.chartData.length)
-      return {
-        labels: this.labels,
-        datasets: this.chartData.map((item, index) => {
-          return {
-            label: this.items[index],
-            data: item,
-            backgroundColor: graphSeries[index].fillColor,
-            borderColor: graphSeries[index].strokeColor,
-            borderWidth: 1,
-            datalabels: {
-              color: index === 0 ? 'white' : 'black',
-              font: {
-                size: '12',
-                weight: index === 0 ? 'bold' : 'normal'
-              }
-            }
-          }
-        })
-      }
-    },
-    displayOption() {
-      const options: Chart.ChartOptions = {
-        tooltips: {
-          enabled: false
-        },
-        responsive: false,
-        maintainAspectRatio: false,
-        legend: {
-          display: false
-        },
-        scales: {
-          xAxes: [
-            {
-              id: 'day',
-              stacked: false,
-              gridLines: {
-                display: false
-              },
-              ticks: {
-                fontSize: 9,
-                maxTicksLimit: 20,
-                fontColor: '#808080',
-                maxRotation: 0,
-                callback: (label: string) => {
-                  return label.split('/')[1]
-                }
-              }
-              // #2384: If you set "type" to "time", make sure that the bars at both ends are not hidden.
-              // #2384: typeをtimeに設定する時はグラフの両端が見切れないか確認してください
-            },
-            {
-              id: 'month',
-              stacked: false,
-              gridLines: {
-                drawOnChartArea: false,
-                drawTicks: true,
-                drawBorder: false,
-                tickMarkLength: 10
-              },
-              ticks: {
-                fontSize: 11,
-                fontColor: '#808080',
-                padding: 3,
-                fontStyle: 'bold'
-              },
-              type: 'time',
-              time: {
-                unit: 'month',
-                parser: 'M/D',
-                displayFormats: {
-                  month: 'MMM'
-                }
-              }
-            }
-          ],
-          yAxes: [
-            {
-              stacked: false,
-              gridLines: {
-                display: true,
-                color: '#E5E5E5'
-              },
-              ticks: {
-                suggestedMin: 0,
-                suggestedMax: this.scaledTicksYAxisMax,
-                maxTicksLimit: 8,
-                fontColor: '#808080'
-              }
-            }
-          ]
-        }
-      }
-      if (this.$route.query.ogp === 'true') {
-        Object.assign(options, { animation: { duration: 0 } })
-      }
-      return options
-    },
-    displayDataHeader() {
-      let n = 0
-      let max = 0
-      for (const i in this.displayData.datasets[0].data) {
-        const current =
-          this.displayData.datasets[0].data[i] +
-          this.displayData.datasets[1].data[i]
-        if (current > max) {
-          max = current
-          n = Number(i)
-        }
-      }
-      return {
-        labels: ['2020/1/1'],
-        datasets: [
-          {
-            data: [this.displayData.datasets[0].data[n]],
-            backgroundColor: 'transparent',
-            borderWidth: 0,
-            datalabels: {
-              font: {
-                size: '0'
-              }
-            }
-          },
-          {
-            data: [this.displayData.datasets[1].data[n]],
-            backgroundColor: 'transparent',
-            borderWidth: 0,
-            datalabels: {
-              font: {
-                size: '0'
-              }
-            }
-          }
-        ]
-      }
-    },
-    displayOptionHeader() {
-      const options: Chart.ChartOptions = {
-        responsive: false,
-        maintainAspectRatio: false,
-        legend: {
-          display: false
-        },
-        tooltips: {
-          enabled: false
-        },
-        scales: {
-          xAxes: [
-            {
-              id: 'day',
-              stacked: false,
-              gridLines: {
-                display: false
-              },
-              ticks: {
-                fontSize: 9,
-                maxTicksLimit: 20,
-                fontColor: 'transparent',
-                maxRotation: 0,
-                minRotation: 0,
-                callback: (label: string) => {
-                  return label.split('/')[1]
-                }
-              }
-            },
-            {
-              id: 'month',
-              stacked: false,
-              gridLines: {
-                drawOnChartArea: false,
-                drawTicks: false, // true -> false
-                drawBorder: false,
-                tickMarkLength: 10
-              },
-              ticks: {
-                fontSize: 11,
-                fontColor: 'transparent', // #808080
-                padding: 13, // 3 + 10(tickMarkLength)
-                fontStyle: 'bold',
-                callback: (label: string) => {
-                  const monthStringArry = [
-                    'Jan',
-                    'Feb',
-                    'Mar',
-                    'Apr',
-                    'May',
-                    'Jun',
-                    'Jul',
-                    'Aug',
-                    'Sep',
-                    'Oct',
-                    'Nov',
-                    'Dec'
-                  ]
-                  const month = monthStringArry.indexOf(label.split(' ')[0]) + 1
-                  return month + '月'
-                }
-              },
-              type: 'time',
-              time: {
-                unit: 'month'
-              }
-            }
-          ],
-          yAxes: [
-            {
-              stacked: false,
-              gridLines: {
-                display: true,
-                drawOnChartArea: false,
-                color: '#E5E5E5' // #E5E5E5
-              },
-              ticks: {
-                suggestedMin: 0,
-                maxTicksLimit: 8,
-                fontColor: '#808080', // #808080
-                suggestedMax: this.scaledTicksYAxisMax
-              }
-            }
-          ]
-        },
-        animation: { duration: 0 }
-      }
-      return options
-    },
-    scaledTicksYAxisMax() {
-      let max = 0
-      for (const i in this.chartData[0]) {
-        max = Math.max(max, this.chartData[0][i] + this.chartData[1][i])
-      }
-      return max
-    },
-    graphWidth() {
-      const window = this.chartWidth ? this.chartWidth : 0
-      const calc = this.displayData.labels!.length * 100
-      return Math.max(window, calc)
-    }
-  },
-  methods: {
-    onClickLegend(i) {
-      this.displayLegends[i] = !this.displayLegends[i]
-      this.displayLegends = this.displayLegends.slice()
-    },
-    cumulative(array: number[]): number[] {
-      const cumulativeArray: number[] = []
-      let patSum = 0
-      array.forEach(d => {
-        patSum += d
-        cumulativeArray.push(patSum)
-      })
-      return cumulativeArray
-    },
-    sum(array: number[]): number {
-      return array.reduce((acc, cur) => {
-        return acc + cur
-      })
-    },
-    pickLastNumber(chartDataArray: number[][]) {
-      return chartDataArray.map(array => {
-        return array[array.length - 1]
-      })
-    },
-    cumulativeSum(chartDataArray: number[][]) {
-      return chartDataArray.map(array => {
-        return array.reduce((acc, cur) => {
-          return acc + cur
-        })
-      })
-    },
-    eachArraySum(chartDataArray: number[][]) {
-      const sumArray: number[] = []
-      for (let i = 0; i < chartDataArray[0].length; i++) {
-        sumArray.push(chartDataArray[0][i] + chartDataArray[1][i])
-      }
-      return sumArray
-    }
-  },
-  mounted() {
-    if (this.$el) {
-      this.chartWidth =
-        ((this.$el!.clientWidth - 22 * 2 - 38) / 60) * this.labels.length + 38
-      this.chartWidth = Math.max(
-        this.$el!.clientWidth - 22 * 2,
-        this.chartWidth
-      )
-    }
-    const barChart = this.$refs.barChart as Vue
-    const barElement = barChart.$el
-    const canvas = barElement.querySelector('canvas')
-    const labelledbyId = `${this.titleId}-graph`
-
-    // スクロールする幅が大きい分には問題ないので大きめにした 本来は適切な値を計算すべき
-    canvas!.parentElement!.parentElement!.parentElement!.scrollLeft! = 2000
-
-    if (canvas) {
-      canvas.setAttribute('role', 'img')
-      canvas.setAttribute('aria-labelledby', labelledbyId)
-    }
-  }
-}
-
-export default Vue.extend(options)
-</script>
-
-<style module lang="scss">
-.Graph {
-  &Legend {
-    text-align: center;
-    list-style: none;
-    padding: 0 !important;
-    li {
-      display: inline-block;
-      margin: 0 3px;
-      div {
-        height: 12px;
-        margin: 2px 4px;
-        width: 40px;
-        display: inline-block;
-        vertical-align: middle;
-        border-width: 1px;
-        border-style: solid;
-      }
-      button {
-        color: $gray-3;
-        @include font-size(12);
-      }
-    }
-  }
-}
-</style>
